Use alternative format string for uint64_t to make printable on 32-bit.
git-svn-id: https://svn.macosforge.org/repository/libdispatch/trunk@117 5710d607-3af0-45f8-8f96-4508d4f60227

+++ testing/dispatch_starfish.c
@@ -61,7 +61,7 @@ collect(void *context __attribute__((unused)))
 
 	printf("lap: %ld\n", lap_count_down);
 	printf("count: %lu\n", COUNT);
-	printf("delta: %llu ns\n", delta);
+	printf("delta: %ju ns\n", (uintmax_t)delta);
 	printf("math: %Lf ns / lap\n", math);
 
 	for (i = 0; i < COUNT; i++) {
